build(lint): add ruff + CI + Python 3.12-3.14 test matrix
- ruff as lint+format (pyproject.toml [tool.ruff]), scoped to everything
  except strategies/ (migrating out per docs/plans/refactor_librae_decouple.md,
  not worth cleaning up pre-move). dev extra pins ruff==0.16.0.
- .githooks/pre-commit mirrors CI on staged files; activate with
  `git config core.hooksPath .githooks`.
- .github/workflows/lint.yml (new): ruff check + format --check.
- .github/workflows/us-live-tests.yml (new): ib_async tests had a pytest
  marker but no CI job.
- core-tests.yml / tw-live-tests.yml / lint.yml switched from bare pip to
  `astral-sh/setup-uv` + `uv sync --frozen`, and from Python 3.11 (below
  the project's own requires-python>=3.12) to a 3.12/3.13/3.14 matrix,
  verified locally (full suite + tw_live/us_live-marked tests) before
  declaring support.
- LICENSE added (pyproject.toml already declared MIT; no file existed).
- Fixed the actual lint/format findings directly across librae/,
  notifications/, orchestration/, db/, brokers/, app/, scripts/, tests/
  (import sorting, datetime.UTC, missing TYPE_CHECKING imports, a dead
  if/else branch in ibkr_adapter.py's _parse_dt, zip(strict=True), unused
  unpacked variables, etc.) rather than suppressing — except
  librae/__init__.py, which keeps a per-file-ignore for its deliberately
  category-grouped imports/__all__.

569 tests passing on py3.12/3.13/3.14.

+++ librae/core/metrics.py
@@ -11,17 +11,19 @@ density (see _infer_annual_periods) so intraday timeframes annualize correctly.
 annual_periods (trading days/year, e.g. 365 for crypto, 252 for TW) is only a
 fallback for when density can't be inferred (<2 bars).
 """
+
 from __future__ import annotations
 
 import logging
+from collections.abc import Callable, Sequence
 from datetime import datetime
-from typing import TYPE_CHECKING, Callable, Sequence
+from typing import TYPE_CHECKING
 
 import numpy as np
 import pandas as pd
 
 if TYPE_CHECKING:
-    from librae.backtest.schema import StrategyMetrics
+    from librae.backtest.schema import OrderEventRecord, StrategyMetrics
     from librae.core.executor import TradePnL
 
 from librae.core import EPSILON
@@ -46,7 +48,7 @@ def _infer_annual_periods(index: pd.DatetimeIndex, fallback: int) -> int:
     if span_seconds <= 0:
         return fallback
     span_years = span_seconds / SECONDS_PER_YEAR
-    return max(1, int(round(len(index) / span_years)))
+    return max(1, round(len(index) / span_years))
 
 
 def compute_all(
@@ -84,6 +86,7 @@ def compute_all(
     # WHY: lazy imports — quantstats pulls in matplotlib/scipy (~1-3s),
     # deferred so `import librae` stays fast.
     import quantstats as qs
+
     from librae.backtest.schema import StrategyMetrics
 
     if not equity_values:
@@ -137,8 +140,7 @@ def compute_all(
     # WHY: profit_factor undefined when no losses (all wins) — return None,
     # not 0.0 which misleadingly suggests worst performance.
     profit_factor = (
-        float(wins.sum() / (losses_abs.sum() + EPSILON))
-        if len(losses_abs) > 0 else None
+        float(wins.sum() / (losses_abs.sum() + EPSILON)) if len(losses_abs) > 0 else None
     )
 
     _comp = _safe_qs(qs.stats.comp, returns)
